bug fix non-target lesion NE
In the previous code, if not all non-target lesions are assessed for a visit, non-target responses are summarized from assessed ones (not an automatic NE). This improvement keeps this behavior but includes a flag (all_nt_assessed) to indicate whether all nontarget lesions are assessed in the output data frame. When compared to the RECIST form, this flag may be used or ignored.